add filter fix
