+++ src/pages/NoticesPage/NoticesPage.jsx
@@ -12,63 +12,25 @@ import {
   getAllNotices,
   selectNoticesTotalPages,
   selectNoticesPage,
-  selectNoticesLoading,
 } from "../../redux/notices/notices-selectors";
 import NoticesSearch from "../../modules/Notices/NoticesSearch/NoticesSearch";
 import NoticesCategoriesNav from "../../modules/Notices/NoticesCategoriesNav/NoticesCategoriesNav";
 import PaginationNotices from "../../shared/components/Pagination/PaginationNotices";
-import NoNoticesFound from "../../modules/Notices/NoNoticesFound/NoNoticesFound";
 
 import css from "../NoticesPage/NoticesPage.module.css";
 import ScrollButton from "../../shared/components/ScrollButton/ScrollButton";
-import {
-  selectAgeFilter,
-  selectGenderFilter,
-} from "../../redux/filter/filter-selectors";
+
 import { keyboard } from "@testing-library/user-event/dist/keyboard";
 
 const NoticesPage = () => {
   const dispatch = useDispatch();
   const notices = useSelector(getAllNotices);
-  const isLoading = useSelector(selectNoticesLoading);
   const [ownCurrentPage, setOwnCurrentPage] = useState(1);
   const [favoriteCurrentPage, setFavoriteCurrentPage] = useState(1);
   const totalPages = useSelector(selectNoticesTotalPages);
   const currentPage = useSelector(selectNoticesPage);
   const location = useLocation();
   const currentCategory = location.pathname.split("/")[2];
-  const ageFilter = useSelector(selectAgeFilter);
-  const genderFilter = useSelector(selectGenderFilter);
-  const [filteredNotices, setFilteredNotices] = useState([]);
-  useEffect(() => {
-    const filtered = notices.filter((notice) => {
-      const isGenderMatch =
-        genderFilter.length === 0 || genderFilter.includes(notice.sex);
-      const isAgeMatch =
-        ageFilter.length === 0 ||
-        ageFilter.some((ageRange) => {
-          if (ageRange === "0-1") {
-            return (
-              parseInt(notice.date.split(".")[2]) >=
-              new Date().getFullYear() - 1
-            );
-          } else if (ageRange === "1-2") {
-            const birthYear = parseInt(notice.date.split(".")[2]);
-            const currentYear = new Date().getFullYear();
-            return birthYear >= currentYear - 2 && birthYear < currentYear - 1;
-          } else if (ageRange === "2+") {
-            return (
-              parseInt(notice.date.split(".")[2]) <=
-              new Date().getFullYear() - 2
-            );
-          }
-          return false;
-        });
-      return isGenderMatch && isAgeMatch;
-    });
-
-    setFilteredNotices(filtered);
-  }, [notices, genderFilter, ageFilter]);
 
   useEffect(() => {
     if (currentCategory === "sell") {
@@ -156,12 +118,7 @@ const NoticesPage = () => {
         onOwnClick={handleOwnClick}
         onFavoriteClick={handleFavoriteClick}
       />
-
-      {filteredNotices.length === 0 && !isLoading ? (
-        <NoNoticesFound title={"notices"} />
-      ) : (
-        <Outlet />
-      )}
+      {notices && <Outlet />}
       <PaginationNotices
         currentPage={currentPage}
         totalPages={totalPages}
